data-loader/requirements.txt
pandas
numpy
requests
click
bs4
html-table-parser-python3
azure-storage-file-datalake
lxml